Refactor admin RBAC e2e helpers
